script/requirements.txt
requests~=2.26.0
PyYAML~=5.4.1
bs4~=0.0.1
beautifulsoup4~=4.9.3